# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для платформы для мастеров ручной работы",
        "title": "CraftPlace",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/shops": {
            "get": {
                "description": "Возвращает список магазинов с возможностью фильтрации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Поиск"
                ],
                "summary": "Получить магазины",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию магазина",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "default": "00000000-0000-0000-0000-000000000000",
                        "description": "Фильтр по ID пользователя",
                        "name": "id_user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqresp.ShopResponse"
                            }
                        }
                    }
                }
            }
        },
        "/shops/{id_shop}": {
            "get": {
                "description": "Возвращает информацию о магазине по его идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Поиск"
                ],
                "summary": "Получить магазин по ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID магазина",
                        "name": "id_shop",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о магазине",
                        "schema": {
                            "$ref": "#/definitions/reqresp.ShopResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID магазина",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Магазин не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/update-login": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Изменяет логин текущего авторизованного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пользователь"
                ],
                "summary": "Обновить логин пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления логина",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqresp.UpdateLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/update-password": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Изменяет пароль текущего авторизованного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пользователь"
                ],
                "summary": "Обновить пароль пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления пароля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqresp.UpdateUserPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/user/user-shops": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет данные указанного магазина пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Магазины"
                ],
                "summary": "Обновить магазин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления магазина",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqresp.UpdateShopRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Магазин успешно обновлен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый магазин для текущего авторизованного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Магазины"
                ],
                "summary": "Добавить магазин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные нового магазина",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqresp.AddShopRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Магазин успешно создан",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет магазин пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Магазины"
                ],
                "summary": "Удалить магазин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для удаления магазина",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqresp.DeleteShopRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Магазин успешно удален",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/{id_user}": {
            "get": {
                "description": "Возвращает информацию о пользователе по его идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пользователь"
                ],
                "summary": "Получить пользователя по ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID пользователя",
                        "name": "id_user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "schema": {
                            "$ref": "#/definitions/reqresp.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID пользователя",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "reqresp.AddShopRequest": {
            "type": "object",
            "required": [
                "description",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Магазин сережек"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Звезды"
                }
            }
        },
...
        "reqresp.DeleteShopRequest": {
            "type": "object",
            "properties": {
                "id_shop": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                }
            }
        },
...
        "reqresp.ShopResponse": {
            "type": "object",
            "required": [
                "description",
                "userID"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучший магазин сережек"
                },
                "id_shop": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "title": {
                    "type": "string",
                    "example": "Eco"
                },
                "userID": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                }
            }
        },
        "reqresp.UpdateLoginRequest": {
            "type": "object",
            "required": [
                "login"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "uname"
                }
            }
        },
...
        "reqresp.UpdateShopRequest": {
            "type": "object",
            "required": [
                "description",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучший магазин сережек"
                },
                "id_shop": {
                    "type": "string",
                    "example": "bb2e8400-e29b-41d4-a716-446655442222"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Лучшие звезды"
                }
            }
        },
        "reqresp.UpdateUserPasswordRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 4,
                    "example": "12345678"
                }
            }
        },
        "reqresp.UserResponse": {
            "type": "object",
            "required": [
                "login"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "uname"
                }
            }
        }
    }
}